ImplÃ©mentation des symboles f(un|try|)lockfile pour libpthread Note: En raison de contrainte sur la taille d'une structure (un void* + deux int pour rÃ©aliser un mutex rÃ©cursif) j'ai changÃ© marcel->number en int au lieu de long
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@12162 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ marcel/include/marcel_descr.h
@@ -114,7 +114,7 @@ struct marcel_task {
 	int static_stack;
 	long initial_sp, depl;
 	char name[MARCEL_MAXNAMESIZE];
-	long number;
+	int number;
 	unsigned long remaining_sleep_time;
 	struct ma_timer_list *timer;
 	unsigned not_migratable, not_deviatable;
